Archivo de configuracion de rutas y limpieza de codigo en componente de calendario
