+++ dental-landing/src/Render/pages/index.ts
@@ -1,0 +1,35 @@
+export { default as AppointmentManagement } from './AppointmentManagement/AppointmentManagement';
+export * from './AppointmentManagement/Calendar/Calendar';
+export * from './AppointmentManagement/Calendar/EventContent/EventContent';
+export * from './AppointmentManagement/Calendar/UpdateAppointmentForm/UpdateAppointmentForm';
+export * from './CreateAppointment/AddAppointment/DateTimeInput/DateTimeInput';
+export * from './CreateAppointment/AddAppointment/SearchPatientInput/SearchPatientInput';
+export * from './CreateAppointment/AddAppointment/SelectInput/SelectInput';
+export * from './CreateAppointment/AddAppointment/TimeInput/TimeInput';
+export * from './CreateAppointment/AddAppointment';
+export { default as CreateAppointment } from './CreateAppointment/CreateAppointment';
+export { default as AppointmentList } from './Dashboard/AppointmentList/AppointmentList';
+export * from './Dashboard/Dashboard';
+export * from './Landing/Landing';
+export * from './Login/ForgotPassword';
+export * from './Login/Login';
+export { default as ResetPassword } from './Login/ResetPassword';
+export { default as NotFound } from './NotFound/NotFound';
+export { default as MedicalHistory } from './PatientManagement/MedicalHistory';
+export * from './PatientManagement/NewPatient';
+export { default as PatientsList } from './PatientManagement/PatientsList';
+export { default as SeeEditMedicalHistory } from './PatientManagement/SeeEditMedicalHistory';
+export * from './PatientManagement/SeeEditPatient';
+export * from './PatientManagement/SeeEditPersonalInfo';
+export { default as Reports } from './Reports/Reports';
+export * from './Reports/Statistics/AppointmentsStatistics/AppointmentsByConfirm/AppointmentsByConfirm';
+export * from './Reports/Statistics/AppointmentsStatistics/AppointmentsByMonthChart/AppointmentsByMonthChart';
+export * from './Reports/Statistics/AppointmentsStatistics/AppointmentsStatistics';
+export * from './Reports/Statistics/PatientsStatistics/PatientsByAgeChart/PatientsByAgeChart';
+export * from './Reports/Statistics/PatientsStatistics/PatientsByGenderChart/PatientsByGenderChart';
+export * from './Reports/Statistics/PatientsStatistics/PatientsByTreatmentChart/PatientsByTreatmentChart';
+export * from './Reports/Statistics/PatientsStatistics/PatientsStatistics';
+export * from './Reports/Statistics';
+export { default as CreateUser } from './UserManagement/CreateUser';
+export * from './UserManagement/EditUserInfo';
+export { default as UsersList } from './UsersManagement/UsersList';
