+++ dental-landing/src/hooks/useScreenWidth.tsx
@@ -1,0 +1,16 @@
+import { useState, useEffect } from "react";
+
+export const useScreenWidth = () => {
+  const [screenWidth, setScreenWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    const handleResize = () => setScreenWidth(window.innerWidth);
+
+    window.addEventListener("resize", handleResize);
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  return screenWidth;
+};
